Refactor the custom code block into a separate module
